+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Added
 * Recipes can spawn parent studies.
 * Study objects can find their own Recipes.
 * `generate` function to spawn studies and immediately schedule them.
+* `generate_all` function to generate studies on multiple symbols in parallel.
 
 Fixed
 -----